requirements.txt
Flask==2.3.3
tensorflow==2.15.0
numpy
pillow
gunicorn